Reload watched (sub)modules automatically
Adds a new step, ReloadModules, which reloads all watched modules except for the init module.

This step needs to know which modules have changed, which is currently only available in the main loops variables (new_changed, last_changed), so I pass this information to the step if it's named "ReloadModules". Other alternatives could be some global state, or passing some general state to all the steps.