Add retry stat on FileNotFound
Solves part of https://github.com/teodorlu/hotload/issues/4

Some editors move the source file before writing to it anew.
This may cause a FileNotFoundError when trying to access the file.
In these cases, we can delay a short time and try again, a limited number of times.
I went for a progressive delay since a short delay often seems to suffice in my experience.

    +---------+--------------+-------------+--------------+
    |         | Post-attempt | Pre-attempt | Total time   |
    | Attempt | delay (ms)*  | delay (ms)  | elapsed (ms) |
    +---------+--------------+-------------+--------------+
    |   1     |      10      |       0     |     0        |
    |   2     |      20      |      10     |    10        |
    |   3     |      30      |      20     |    30        |
    |   4     |       0      |      30     |    60        |
    +---------+--------------+-------------+--------------+
               (*on failure)

The total is the same as 20 ms fixed delay: 20 + 20 + 20 = 60.