Test: shell w/o sleep/poll
Rewrite shell test to avoid `sleep` and polling.

The output from hotload.py is piped to a subshell command group
containing a sequence of `expect` function calls.
These functions assert that the expected strings are observered, in order.

After running the test, we want to tidy up and end the python process. To get its PID we need to use a named pipe, `"$pipe"`.

To avoid infinite "hang", a timeout is set by running `sleep` in parallell with the tests. Again a named pipe is needed, in this case to wait for the first of the jobs (tests, sleep) to finish. The main process reads from FIFO pipe `$done_chan`, blocking until a message is received.