Test: traps galore!
= Signal madness =

Make sure all child processes are killed when script ends, both when successful and otherwise.

Put cleanup code in a single top-level function, `end()`. So far so good.

Send kill signal to all children by getting pgid (process group id) of all jobs.

The script should quit with the correct exit code, so we have to set `trap 'exit $?' TERM`.

We cannot send SIGINT to the asynchronous commands (it's ignored when job control is disabled),
so TERM is sent, and `trap 'trap - TERM; kill -s INT $$' TERM` is set for the parent process.