Test: PID armageddon!
Kill each PID decentant.

GitHub runs the script in the same process group as the runner, so killing the process group is problematic.

This commit kills every decendant PID instead.

Not sure if the order matters, so to be on the safe side the youngest PIDs are printed first.

I.e. we call `kill GRANDCHILD1 GRANDCHILD2 CHILD1 CHILD2 PARENT`

The function prints lines of:

    ...
    GREAT ** N  GRANDCHILDREN  (N >= 2)
    GREAT GRANDCHILDREN
    GRANDCHILDREN
    CHILDREN
    PARENT (`root`)