Test2 (#11)
Shell tests avoiding `sleep` and polling.

The output from hotload.py is piped to a subshell command group 
containing a sequence of `expect` function calls.
These functions assert that the expected strings are observed, in order. (Ignoring other output interspersed.)

After running the test, we want to tidy up and end the python process. To get its PID we need to use a named pipe, `"$pipe"`.

To avoid infinite "hang", a timeout is set by running `sleep` in parallel with the tests. Again a named pipe is needed, in this case to wait for the first of the jobs (tests, sleep) to finish. The main process reads from FIFO pipe `$done_chan`, blocking until a message is received.

The test runs pretty fast: https://github.com/teodorlu/hotload/runs/7297455043

```
2022-07-12T07:58:27.8053635Z ##[group]Run sh -x .github/workflows/test.sh
...
2022-07-12T07:58:28.2685395Z Tests OK!
```
=> 0.463 seconds. About 0.09s on my laptop. Remember, this includes starting python and several other processes (sed etc.).

### Traps galore — Signal madness 

We want to clean up all processes on potential errors.

Make sure all child processes are killed when script ends, both when successful and otherwise.

Cleanup code is put in a single top-level function, `end()`.

The script should quit with the correct exit code, so we have to set `trap 'exit $?' TERM`.

We cannot send SIGINT to the asynchronous commands (it's ignored when job control is disabled),
so TERM is sent, and `trap 'trap - TERM; kill -s INT $$' TERM` is set for the parent process.

### PID armageddon! 

Kill each PID decentant.

GitHub runs the script in the same process group as the runner, so killing the process group is problematic.

This commit kills every descendant PID instead.

Not sure if the order matters, so to be on the safe side the youngest PIDs are printed first.

I.e. we call `kill GRANDCHILD1 GRANDCHILD2 CHILD1 CHILD2 PARENT`

The function prints lines of:

    ...
    GREAT ** N  GRANDCHILDREN  (N >= 2)
    GREAT GRANDCHILDREN
    GRANDCHILDREN
    CHILDREN
    PARENT (`root`)

#### kill $pid and cleanup fifos

On successful runs, $pid was not terminated, getting parent PID 1.

Add EXIT trap to terminate $pid, and also delete FIFOs.

The EXIT trap is called on all exits in Bash, including SIGTERM.

Other shells runs EXIT trap only in the absence of signals.

Keeping this generic is too much work, the shebang reflects Bash requirement.